refactor: decompose PostgresObservabilityStore (499→48+196+253) into partials
Split by CQRS alignment: shell+connection, write operations, read queries.
