+++ src/Content/Infrastructure/Infrastructure.Observability/Persistence/PostgresObservabilityStore.cs
@@ -1,9 +1,6 @@
-using System.Text.Json;
 using Application.AI.Common.Interfaces;
-using Domain.AI.Observability.Models;
 using Microsoft.Extensions.Logging;
 using Npgsql;
-using NpgsqlTypes;
 
 namespace Infrastructure.Observability.Persistence;
 
@@ -13,7 +10,7 @@ namespace Infrastructure.Observability.Persistence;
 /// for non-critical writes (audit, safety) to avoid blocking agent turns.
 /// Read methods return empty collections on failure to maintain resilience.
 /// </summary>
-public sealed class PostgresObservabilityStore : IObservabilityStore, IDisposable
+public sealed partial class PostgresObservabilityStore : IObservabilityStore, IDisposable
 {
     private readonly NpgsqlDataSource _dataSource;
     private readonly ILogger<PostgresObservabilityStore> _logger;
@@ -25,457 +22,6 @@ public sealed class PostgresObservabilityStore : IObservabilityStore, IDisposabl
         _dataSource = NpgsqlDataSource.Create(connectionString);
         _logger = logger;
     }
-
-    // ── Sessions ─────────────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task<Guid> StartSessionAsync(
-        string conversationId, string agentName, string? model,
-        CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            INSERT INTO sessions (conversation_id, agent_name, model)
-            VALUES ($1, $2, $3)
-            ON CONFLICT (conversation_id) DO UPDATE SET started_at = NOW()
-            RETURNING id
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(conversationId);
-            cmd.Parameters.AddWithValue(agentName);
-            cmd.Parameters.AddWithValue(model ?? (object)DBNull.Value);
-
-            var result = await cmd.ExecuteScalarAsync(cancellationToken);
-            return result is Guid id ? id : Guid.Empty;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to start session {ConversationId}", conversationId);
-            return Guid.Empty;
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task EndSessionAsync(
-        Guid sessionId, string status, string? errorMessage,
-        CancellationToken cancellationToken = default)
-    {
-        if (sessionId == Guid.Empty) return;
-
-        const string sql = """
-            UPDATE sessions
-            SET ended_at = NOW(),
-                duration_ms = EXTRACT(EPOCH FROM (NOW() - started_at))::INT * 1000,
-                status = $2,
-                error_message = $3
-            WHERE id = $1
-            """;
-
-        await ExecuteNonQuerySafe(sql, cancellationToken,
-            sessionId, status, errorMessage ?? (object)DBNull.Value);
-    }
-
-    /// <inheritdoc />
-    public async Task UpdateSessionMetricsAsync(
-        Guid sessionId, int turnCount, int toolCallCount, int subagentCount,
-        int totalInputTokens, int totalOutputTokens, int totalCacheRead,
-        int totalCacheWrite, decimal totalCostUsd, decimal cacheHitRate,
-        string? model = null, CancellationToken cancellationToken = default)
-    {
-        if (sessionId == Guid.Empty) return;
-
-        const string sql = """
-            UPDATE sessions
-            SET turn_count = $2,
-                tool_call_count = $3,
-                subagent_count = $4,
-                total_input_tokens = $5,
-                total_output_tokens = $6,
-                total_cache_read = $7,
-                total_cache_write = $8,
-                total_cost_usd = $9,
-                cache_hit_rate = $10,
-                model = COALESCE($11, model),
-                duration_ms = EXTRACT(EPOCH FROM (NOW() - started_at))::INT * 1000
-            WHERE id = $1
-            """;
-
-        await ExecuteNonQuerySafe(sql, cancellationToken,
-            sessionId, turnCount, toolCallCount, subagentCount,
-            totalInputTokens, totalOutputTokens, totalCacheRead,
-            totalCacheWrite, totalCostUsd, cacheHitRate,
-            (object?)model ?? DBNull.Value);
-    }
-
-    // ── Messages ─────────────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task<Guid> RecordMessageAsync(
-        Guid sessionId, int turnIndex, string role, string? source,
-        string? contentPreview, string? model, int inputTokens, int outputTokens,
-        int cacheRead, int cacheWrite, decimal costUsd, decimal cacheHitPct,
-        string[]? toolNames = null, CancellationToken cancellationToken = default)
-    {
-        if (sessionId == Guid.Empty) return Guid.Empty;
-
-        const string sql = """
-            INSERT INTO session_messages
-                (session_id, turn_index, role, source, content_preview, model,
-                 input_tokens, output_tokens, cache_read, cache_write,
-                 cost_usd, cache_hit_pct, tool_names)
-            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13)
-            RETURNING id
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(sessionId);
-            cmd.Parameters.AddWithValue(turnIndex);
-            cmd.Parameters.AddWithValue(role);
-            cmd.Parameters.AddWithValue(source ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue(contentPreview ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue(model ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue(inputTokens);
-            cmd.Parameters.AddWithValue(outputTokens);
-            cmd.Parameters.AddWithValue(cacheRead);
-            cmd.Parameters.AddWithValue(cacheWrite);
-            cmd.Parameters.AddWithValue(costUsd);
-            cmd.Parameters.AddWithValue(cacheHitPct);
-            cmd.Parameters.Add(new NpgsqlParameter
-            {
-                Value = toolNames ?? (object)DBNull.Value,
-                NpgsqlDbType = NpgsqlDbType.Array | NpgsqlDbType.Text
-            });
-
-            var result = await cmd.ExecuteScalarAsync(cancellationToken);
-            return result is Guid id ? id : Guid.Empty;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to record message for session {SessionId}", sessionId);
-            return Guid.Empty;
-        }
-    }
-
-    // ── Tool Executions ──────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task RecordToolExecutionAsync(
-        Guid sessionId, Guid? messageId, string toolName, string toolSource,
-        int durationMs, string status, string? errorType,
-        int? resultSize, CancellationToken cancellationToken = default)
-    {
-        if (sessionId == Guid.Empty) return;
-
-        const string sql = """
-            INSERT INTO tool_executions
-                (session_id, message_id, tool_name, tool_source,
-                 duration_ms, status, error_type, result_size)
-            VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
-            """;
-
-        await ExecuteNonQuerySafe(sql, cancellationToken,
-            sessionId,
-            messageId.HasValue && messageId.Value != Guid.Empty ? messageId.Value : DBNull.Value,
-            toolName, toolSource, durationMs, status,
-            errorType ?? (object)DBNull.Value,
-            resultSize.HasValue ? resultSize.Value : DBNull.Value);
-    }
-
-    // ── Safety Events ────────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task RecordSafetyEventAsync(
-        Guid sessionId, string phase, string outcome, string? category,
-        int? severity, string? filterName,
-        CancellationToken cancellationToken = default)
-    {
-        if (sessionId == Guid.Empty) return;
-
-        const string sql = """
-            INSERT INTO safety_events
-                (session_id, phase, outcome, category, severity, filter_name)
-            VALUES ($1, $2, $3, $4, $5, $6)
-            """;
-
-        await ExecuteNonQuerySafe(sql, cancellationToken,
-            sessionId, phase, outcome,
-            category ?? (object)DBNull.Value,
-            severity.HasValue ? severity.Value : DBNull.Value,
-            filterName ?? (object)DBNull.Value);
-    }
-
-    // ── Audit ────────────────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task RecordAuditAsync(
-        string operation, string source, Guid? sessionId,
-        Dictionary<string, object>? metadata,
-        CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            INSERT INTO audit_log (operation, source, session_id, metadata)
-            VALUES ($1, $2, $3, $4::jsonb)
-            """;
-
-        await ExecuteNonQuerySafe(sql, cancellationToken,
-            operation, source,
-            sessionId.HasValue && sessionId.Value != Guid.Empty ? sessionId.Value : DBNull.Value,
-            metadata is not null ? JsonSerializer.Serialize(metadata) : DBNull.Value);
-    }
-
-    // ── Reads ────────────────────────────────────────────────────────────
-
-    /// <inheritdoc />
-    public async Task<IReadOnlyList<SessionRecord>> GetSessionsAsync(
-        int limit = 50, int offset = 0, string? status = null,
-        DateTimeOffset? since = null, DateTimeOffset? until = null,
-        CancellationToken cancellationToken = default)
-    {
-        const string columns = """
-            SELECT id, conversation_id, agent_name, model, started_at, ended_at, duration_ms,
-                   turn_count, tool_call_count, subagent_count, total_input_tokens, total_output_tokens,
-                   total_cache_read, total_cache_write, total_cost_usd, cache_hit_rate,
-                   status, error_message, created_at
-            FROM sessions
-            """;
-
-        var clauses = new List<string>();
-        var paramIndex = 3;
-
-        if (!string.IsNullOrWhiteSpace(status))
-            clauses.Add($"status = ${paramIndex++}");
-        if (since.HasValue)
-            clauses.Add($"started_at >= ${paramIndex++}");
-        if (until.HasValue)
-            clauses.Add($"started_at < ${paramIndex++}");
-
-        var where = clauses.Count > 0 ? "WHERE " + string.Join(" AND ", clauses) : "";
-        var sql = $"{columns} {where} ORDER BY started_at DESC LIMIT $1 OFFSET $2";
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(limit);
-            cmd.Parameters.AddWithValue(offset);
-            if (!string.IsNullOrWhiteSpace(status))
-                cmd.Parameters.AddWithValue(status!);
-            if (since.HasValue)
-                cmd.Parameters.AddWithValue(since.Value);
-            if (until.HasValue)
-                cmd.Parameters.AddWithValue(until.Value);
-
-            await using var reader = await cmd.ExecuteReaderAsync(cancellationToken);
-            var results = new List<SessionRecord>();
-            while (await reader.ReadAsync(cancellationToken))
-                results.Add(ReadSessionRecord(reader));
-
-            return results;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to read sessions (limit={Limit}, offset={Offset})", limit, offset);
-            return Array.Empty<SessionRecord>();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<SessionRecord?> GetSessionByIdAsync(
-        Guid sessionId, CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            SELECT id, conversation_id, agent_name, model, started_at, ended_at, duration_ms,
-                   turn_count, tool_call_count, subagent_count, total_input_tokens, total_output_tokens,
-                   total_cache_read, total_cache_write, total_cost_usd, cache_hit_rate,
-                   status, error_message, created_at
-            FROM sessions
-            WHERE id = $1
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(sessionId);
-
-            await using var reader = await cmd.ExecuteReaderAsync(cancellationToken);
-            return await reader.ReadAsync(cancellationToken)
-                ? ReadSessionRecord(reader)
-                : null;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to read session {SessionId}", sessionId);
-            return null;
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<IReadOnlyList<SessionMessageRecord>> GetSessionMessagesAsync(
-        Guid sessionId, CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            SELECT id, session_id, turn_index, role, source, content_preview, model,
-                   input_tokens, output_tokens, cache_read, cache_write, cost_usd, cache_hit_pct,
-                   tool_names, created_at
-            FROM session_messages
-            WHERE session_id = $1
-            ORDER BY turn_index
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(sessionId);
-
-            await using var reader = await cmd.ExecuteReaderAsync(cancellationToken);
-            var results = new List<SessionMessageRecord>();
-            while (await reader.ReadAsync(cancellationToken))
-                results.Add(ReadSessionMessageRecord(reader));
-
-            return results;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to read messages for session {SessionId}", sessionId);
-            return Array.Empty<SessionMessageRecord>();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<IReadOnlyList<ToolExecutionRecord>> GetSessionToolExecutionsAsync(
-        Guid sessionId, CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            SELECT id, session_id, message_id, tool_name, tool_source, duration_ms, status,
-                   error_type, result_size, created_at
-            FROM tool_executions
-            WHERE session_id = $1
-            ORDER BY created_at
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(sessionId);
-
-            await using var reader = await cmd.ExecuteReaderAsync(cancellationToken);
-            var results = new List<ToolExecutionRecord>();
-            while (await reader.ReadAsync(cancellationToken))
-                results.Add(ReadToolExecutionRecord(reader));
-
-            return results;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to read tool executions for session {SessionId}", sessionId);
-            return Array.Empty<ToolExecutionRecord>();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<IReadOnlyList<SafetyEventRecord>> GetSessionSafetyEventsAsync(
-        Guid sessionId, CancellationToken cancellationToken = default)
-    {
-        const string sql = """
-            SELECT id, session_id, phase, outcome, category, severity, filter_name, created_at
-            FROM safety_events
-            WHERE session_id = $1
-            ORDER BY created_at
-            """;
-
-        try
-        {
-            await using var cmd = _dataSource.CreateCommand(sql);
-            cmd.Parameters.AddWithValue(sessionId);
-
-            await using var reader = await cmd.ExecuteReaderAsync(cancellationToken);
-            var results = new List<SafetyEventRecord>();
-            while (await reader.ReadAsync(cancellationToken))
-                results.Add(ReadSafetyEventRecord(reader));
-
-            return results;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to read safety events for session {SessionId}", sessionId);
-            return Array.Empty<SafetyEventRecord>();
-        }
-    }
-
-    // ── Reader Helpers ───────────────────────────────────────────────────
-
-    private static SessionRecord ReadSessionRecord(NpgsqlDataReader reader) => new()
-    {
-        Id = reader.GetGuid(0),
-        ConversationId = reader.GetString(1),
-        AgentName = reader.GetString(2),
-        Model = reader.IsDBNull(3) ? null : reader.GetString(3),
-        StartedAt = reader.GetFieldValue<DateTimeOffset>(4),
-        EndedAt = reader.IsDBNull(5) ? null : reader.GetFieldValue<DateTimeOffset>(5),
-        DurationMs = reader.IsDBNull(6) ? null : reader.GetInt32(6),
-        TurnCount = reader.GetInt32(7),
-        ToolCallCount = reader.GetInt32(8),
-        SubagentCount = reader.GetInt32(9),
-        TotalInputTokens = reader.GetInt32(10),
-        TotalOutputTokens = reader.GetInt32(11),
-        TotalCacheRead = reader.GetInt32(12),
-        TotalCacheWrite = reader.GetInt32(13),
-        TotalCostUsd = reader.GetDecimal(14),
-        CacheHitRate = reader.GetDecimal(15),
-        Status = reader.GetString(16),
-        ErrorMessage = reader.IsDBNull(17) ? null : reader.GetString(17),
-        CreatedAt = reader.GetFieldValue<DateTimeOffset>(18)
-    };
-
-    private static SessionMessageRecord ReadSessionMessageRecord(NpgsqlDataReader reader) => new()
-    {
-        Id = reader.GetGuid(0),
-        SessionId = reader.GetGuid(1),
-        TurnIndex = reader.GetInt32(2),
-        Role = reader.GetString(3),
-        Source = reader.IsDBNull(4) ? null : reader.GetString(4),
-        ContentPreview = reader.IsDBNull(5) ? null : reader.GetString(5),
-        Model = reader.IsDBNull(6) ? null : reader.GetString(6),
-        InputTokens = reader.GetInt32(7),
-        OutputTokens = reader.GetInt32(8),
-        CacheRead = reader.GetInt32(9),
-        CacheWrite = reader.GetInt32(10),
-        CostUsd = reader.GetDecimal(11),
-        CacheHitPct = reader.GetDecimal(12),
-        ToolNames = reader.IsDBNull(13) ? null : reader.GetFieldValue<string[]>(13),
-        CreatedAt = reader.GetFieldValue<DateTimeOffset>(14)
-    };
-
-    private static ToolExecutionRecord ReadToolExecutionRecord(NpgsqlDataReader reader) => new()
-    {
-        Id = reader.GetGuid(0),
-        SessionId = reader.GetGuid(1),
-        MessageId = reader.IsDBNull(2) ? null : reader.GetGuid(2),
-        ToolName = reader.GetString(3),
-        ToolSource = reader.IsDBNull(4) ? null : reader.GetString(4),
-        DurationMs = reader.IsDBNull(5) ? null : reader.GetInt32(5),
-        Status = reader.GetString(6),
-        ErrorType = reader.IsDBNull(7) ? null : reader.GetString(7),
-        ResultSize = reader.IsDBNull(8) ? null : reader.GetInt32(8),
-        CreatedAt = reader.GetFieldValue<DateTimeOffset>(9)
-    };
-
-    private static SafetyEventRecord ReadSafetyEventRecord(NpgsqlDataReader reader) => new()
-    {
-        Id = reader.GetGuid(0),
-        SessionId = reader.GetGuid(1),
-        Phase = reader.GetString(2),
-        Outcome = reader.GetString(3),
-        Category = reader.IsDBNull(4) ? null : reader.GetString(4),
-        Severity = reader.IsDBNull(5) ? null : reader.GetInt32(5),
-        FilterName = reader.IsDBNull(6) ? null : reader.GetString(6),
-        CreatedAt = reader.GetFieldValue<DateTimeOffset>(7)
-    };
-
-    // ── Write Helpers ────────────────────────────────────────────────────
 
     private async Task ExecuteNonQuerySafe(
         string sql, CancellationToken cancellationToken, params object[] parameters)
